WC-1362: nhaystack.driver: Add support for TLS.
In cases where the Project Haystack server is on the Internet, it is highly
undesirable to use cleartext HTTP in order to access data as it exposes the
data and authentication credentials to anybody who is able to sniff the
traffic.

The following adds a flag to enable use of Transport Layer Security when
connecting to the target server.

Credit: Benny Reese <https://bitbucket.org/bennyreese1/> in
        <https://bitbucket.org/bennyreese1/nhaystack/branch/ax-ssl>
Ported to `n4` branch.