[rt] Add test stations for history upload.
`nhaystack_sender` is a JACE that pushes its history to another
nHaystack instance running on another JACE via `hisWrite` calls.

`nhaystack_receiver` is a JACE that receives the history from
`nhaystack_sender`.

There's no corresponding test in the `nhaystack` tests, rather the test
case is you spin both up, go away for a few minutes, then manually check
if the data that got historised in the sender was pushed successfully to
the receiver.