[rt] nhaystack: Support truly null `BHRef`.
Presently, if you want a "null" BHRef, the value chosen is a literal
`@null`, which is not intuitive for the user.  It would be better to
have the user just leave a field completely blank.

This however causes parsing errors with the `BHRef` class.

The following commit amends `BHRef` so it can support representing a
"null" `HRef`.  That is, a `HRef` that is not presently set to point to
any given Project Haystack entity.